requirements.txt
fastapi
uvicorn
pydantic-settings
python-jose
cryptography
azure-data-tables
openpyxl
httpx
slowapi
structlog
python-multipart
